+++ EDDI/ChangeLog.md
@@ -2,7 +2,7 @@
 
 Full details of the variables available for each noted event, and VoiceAttack integrations, are available in the individual [event pages](https://github.com/EDCD/EDDI/wiki/Events).
 
-### Development
+### 3.1.0-b5
   * Core
     * EDDI now uses EDSM as the primary backend for server data (rather than the no longer maintained server set up by EDDI's original creator)
     * Strings returned from the player journal and from server data have been standardized and should match each other much more closely than they have in the past.